requirements.txt
Flask
python-dotenv==1.1.1
google-generativeai
pillow==10.4.0
PyMuPDF==1.26.4
neo4j==5.28.2
camel-ai==0.2.75
Werkzeug==3.1.3
opencv-python==4.12.0.88
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
gunicorn==21.2.0
psycopg2-binary==2.9.9